data
raw file containing some coordinate points and the local time when a waste bin was fill to certain level (%). The colors of the bins are indicated by black or blue and the size in cu metre of the bin are indicated under the "no" column.  Indication whether the was containers where emptied or not are indicated by TRUE or FALSE